float

+++ bootstrap/Visualisations/bootstrap_oneCat.js
@@ -106,7 +106,7 @@ class bootstrap_oneCat extends visBase {
 
 		// this.largeCISplit = Math.abs(this.populationStatistic - statlist[10000*0.95]);
 		statList.sort(function(a, b){
-			return parseInt(a) - parseInt(b);
+			return parseFloat(a) - parseFloat(b);
 		});
 		let lower_CI = statList[Math.floor(this.numSamples*0.025)];
 		let upper_CI = statList[Math.ceil(this.numSamples*0.975)];
@@ -115,7 +115,7 @@ class bootstrap_oneCat extends visBase {
 	setUpCI(statList){
 		var CISplit = Math.abs(this.populationStatistic - statList[this.numSamples*0.95]);
 		statList.sort(function(a, b){
-			return parseInt(a) - parseInt(b);
+			return parseFloat(a) - parseFloat(b);
 		});
 		let lower_CI = statList[Math.floor(this.numSamples*0.025)];
 		let upper_CI = statList[Math.ceil(this.numSamples*0.975)];
